py: refactoring: restructured profile.ts
